backend/pyproject.toml
[project]
name = "cockpit-authelia-users"
version = "0.1.0"
description = "Cockpit plugin for Authelia user management"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Hat Labs", email = "info@hatlabs.fi"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "argon2-cffi>=23.1.0",
]

[project.scripts]
cockpit-authelia-users-bridge = "cockpit_authelia_users.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=cockpit_authelia_users",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["cockpit_authelia_users"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "SIM117",  # Multiple with statements - often clearer as-is
    "B904",    # raise-without-from-inside-except - not always appropriate
    "E501",    # Line too long - handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py
"tests/**/*.py" = ["ARG"]  # Unused arguments in tests are OK

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["cockpit_authelia_users"]
exclude = ["**/__pycache__", "**/node_modules", ".venv", ".uv"]
pythonVersion = "3.11"
typeCheckingMode = "strict"
reportMissingImports = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
